\name{svdvisual-package}
\alias{svdvisual-package}
\alias{svdvisual}
\docType{package}
\title{
SVD Visualization Package
}
\description{
This package provides a full spectrum of many useful visualization tools associated with SVD related methodology. 
}
\details{
\tabular{ll}{
Package: \tab svdvisual\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-09-23\cr
License: \tab GPL \cr
}
}
\author{
Lingsong Zhang<lingsong@purdue.edu>
Maintainer: Lingsong Zhang<lingsong@purdue.edu>
}
\references{
Zhang, L., Marron, J.S., Shen, H. and Zhu, Z. (2007) Singular Value Decomposition and Its Visualization, Journal of Computational and Graphical Statistics. 
}
\keyword{svd}
\keyword{visualization}
\seealso{
\code{\link{lattice}} 
}
